requirements.txt
tensorflow-gpu==1.13.1
win32gui
opencv-python
imblearn
